feature/LTBENCH-434: added missing README.md for feign-demo, restassured-demo, selenide-demo modules. Renamed packages to clearly represent their modules
