+++ feign-demo/README.md
@@ -1,0 +1,14 @@
+# Test-automation-maven-feign-demo
+
+## How to:
+* Run feign tests
+
+`mvn clean test -Dtest=FeignDemoTest`
+
+* Check the generated report
+
+`build/allure-report/data/attachments/*.html`
+
+### Additional information:
+
+* Package is used for Back-End scenario testing
