+++ feign-demo/src/test/java/com/nortal/test/feign/demo/configuration/TestDemoConfiguration.java
@@ -1,6 +1,6 @@
 /**
  * Copyright (c) 2022 Nortal AS
- * <p>
+ *
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -8,10 +8,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- * <p>
+ *
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- * <p>
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -20,17 +20,13 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.nortal.test.demo.configuration;
+package com.nortal.test.feign.demo.configuration;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nortal.test.restassured.RestAssuredConfiguration;
-import org.springframework.context.annotation.Bean;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
+@EnableFeignClients(basePackages = {"com.nortal.test.feign.demo.api"})
 public class TestDemoConfiguration {
-    @Bean
-    public ObjectMapper objectMapper(RestAssuredConfiguration.ObjectMapperProvider objectMapperProvider) {
-        return objectMapperProvider.getObjectMapper();
-    }
 }
