+++ feign-demo/src/test/java/com/nortal/test/feign/demo/glue/TestPageStepDefs.java
@@ -1,6 +1,6 @@
 /**
  * Copyright (c) 2022 Nortal AS
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -8,10 +8,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -20,13 +20,21 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.nortal.test.demo.configuration;
+package com.nortal.test.feign.demo.glue;
 
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Configuration;
+import com.nortal.test.feign.demo.api.CatApi;
+import io.cucumber.java.en.Given;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 
-@Configuration
-@EnableFeignClients(basePackages = {"com.nortal.test.demo.api"})
-public class TestDemoConfiguration {
+public class TestPageStepDefs {
+    @Autowired
+    private CatApi catApi;
+
+    @Given("Api is called")
+    public void openPageInBrowser() {
+        ResponseEntity<CatApi.Fact> response = catApi.getFact();
+    }
+
+
 }
