+++ feign-demo/src/test/resources/application-override.yml
@@ -14,7 +14,7 @@ test-automation:
       parallel:
         enabled: true
         executor-count: 2
-    glue-append: "com.nortal.test.demo.glue"
+    glue-append: "com.nortal.test.feign.demo.glue"
     filter:
       tags: "not @Skip"
   report:
