+++ pom.xml
@@ -74,6 +74,9 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>3.0.0-M7</version>
+                <configuration>
+                <failIfNoSpecifiedTests>false</failIfNoSpecifiedTests>
+                </configuration>
             </plugin>
         </plugins>
     </build>
