+++ restassured-demo/README.md
@@ -1,0 +1,14 @@
+# Test-automation-maven-restassured-demo
+
+## How to:
+* Run Restassured tests
+
+`mvn clean test -Dtest=RestassuredDemoTest`
+
+* Check the generated report
+
+`build/allure-report/data/attachments/*.html`
+
+### Additional information:
+
+* Package is used for Back-End scenario testing
