+++ restassured-demo/src/test/java/com/nortal/test/restassured/demo/configuration/TestDemoConfiguration.java
@@ -1,6 +1,6 @@
 /**
  * Copyright (c) 2022 Nortal AS
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -8,10 +8,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -20,31 +20,17 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.nortal.test.demo.configuration;
+package com.nortal.test.restassured.demo.configuration;
 
-import lombok.Getter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nortal.test.restassured.RestAssuredConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
-@Getter
-@ConstructorBinding
-@ConfigurationProperties(prefix = "test-automation.custom")
-public class TestDemoProperties {
-	/**
-	 * Enable demo mode which highlights elements.
-	 */
-	private final boolean demoMode;
-	/**
-	 * Add sleep between Selenide actions.
-	 */
-	private final boolean lazyExecution;
-	private final Long lazyExecutionTime;
-	private final boolean screenshotsSteps;
-
-	public TestDemoProperties(boolean demoMode, boolean lazyExecution, Long lazyExecutionTime, boolean screenshotsSteps) {
-		this.demoMode = demoMode;
-		this.lazyExecution = lazyExecution;
-		this.lazyExecutionTime = lazyExecutionTime;
-		this.screenshotsSteps = screenshotsSteps;
-	}
+@Configuration
+public class TestDemoConfiguration {
+    @Bean
+    public ObjectMapper objectMapper(RestAssuredConfiguration.ObjectMapperProvider objectMapperProvider) {
+        return objectMapperProvider.getObjectMapper();
+    }
 }
