+++ restassured-demo/src/test/java/com/nortal/test/restassured/demo/glue/TestPageStepDefs.java
@@ -1,6 +1,6 @@
 /**
  * Copyright (c) 2022 Nortal AS
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -8,10 +8,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -20,11 +20,20 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.nortal.test.demo;
+package com.nortal.test.restassured.demo.glue;
 
-import com.nortal.test.TestRunner;
-import org.junit.platform.suite.api.SelectClasspathResource;
+import io.cucumber.java.en.Given;
+import io.restassured.RestAssured;
 
-@SelectClasspathResource("/behavior")
-public class RestassuredDemoTest extends TestRunner {
+public class TestPageStepDefs {
+
+    @Given("Api is called")
+    public void openPageInBrowser() {
+        RestAssured.given()
+                .get("https://catfact.ninja/fact")
+                .then()
+                .statusCode(200);
+    }
+
+
 }
