+++ restassured-demo/src/test/resources/application-override.yml
@@ -10,4 +10,4 @@ test-automation:
   report-name: demo-suite
   environment: local
   cucumber:
-    glue-append: "com.nortal.test.demo.glue"
+    glue-append: "com.nortal.test.restassured.demo.glue"
