+++ selenide-demo/README.md
@@ -1,0 +1,14 @@
+# Test-automation-maven-selenide-demo
+
+## How to:
+* Run Selenide tests
+
+`mvn clean test -Dtest=SelenideDemoTest`
+
+* Check the generated step screenshots
+
+`build/allure-report/data/attachments/`
+
+### Additional information:
+
+* Package is used for Front-End scenario testing
