+++ selenide-demo/src/test/java/com/nortal/test/selenide/demo/glue/TestPageStepDefs.java
@@ -20,11 +20,11 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.nortal.test.demo.glue;
+package com.nortal.test.selenide.demo.glue;
 
 import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.Selenide;
-import com.nortal.test.demo.mediator.UiMediator;
+import com.nortal.test.selenide.demo.mediator.UiMediator;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
